+++ src/models/student/StudentNote.kt
@@ -1,21 +1,27 @@
 package com.progressp.models.student
 
+import com.progressp.config.DbContstants
+import com.progressp.models.user.UsersTable
 import org.jetbrains.exposed.dao.UUIDEntity
 import org.jetbrains.exposed.dao.UUIDEntityClass
 import org.jetbrains.exposed.dao.id.EntityID
 import org.jetbrains.exposed.dao.id.UUIDTable
 import org.jetbrains.exposed.sql.Column
 import org.jetbrains.exposed.sql.ReferenceOption
+import org.jetbrains.exposed.sql.javatime.date
 import org.jetbrains.exposed.sql.javatime.datetime
+import java.time.LocalDate
 import java.time.LocalDateTime
 import java.util.UUID
 
 object StudentsNotesTable : UUIDTable("students_notes") {
+    var instructorId: Column<EntityID<UUID>> = reference(
+        "instructor_id", UsersTable, onDelete = ReferenceOption.CASCADE
+    )
     val studentId: Column<EntityID<UUID>> = reference("student_id", StudentsTable, onDelete = ReferenceOption.CASCADE)
-    val weight: Column<Double> = double("weight")
-    val waist: Column<Double> = double("waist")
-    val arms: Column<Double> = double("arms")
-    val legs: Column<Double> = double("legs")
+    val measurementName: Column<String> = varchar("measurement_name", DbContstants.STRING_LENGTH)
+    val measurementValue: Column<Double> = double("measurement_value")
+    val tookAt: Column<LocalDate> = date("took_at").default(LocalDate.now())
     val updatedAt: Column<LocalDateTime> = datetime("updated_at").default(LocalDateTime.now())
     val createdAt: Column<LocalDateTime> = datetime("created_at").default(LocalDateTime.now())
 }
@@ -23,31 +29,59 @@ object StudentsNotesTable : UUIDTable("students_notes") {
 class StudentNote(id: EntityID<UUID>) : UUIDEntity(id) {
     companion object : UUIDEntityClass<StudentNote>(StudentsNotesTable)
 
+    var instructorId by StudentsNotesTable.instructorId
     var studentId by StudentsNotesTable.studentId
-    var weight by StudentsNotesTable.weight
-    var waist by StudentsNotesTable.waist
-    var arms by StudentsNotesTable.arms
-    var legs by StudentsNotesTable.legs
+    var measurementName by StudentsNotesTable.measurementName
+    var measurementValue by StudentsNotesTable.measurementValue
+    var tookAt by StudentsNotesTable.tookAt
     var updatedAt by StudentsNotesTable.updatedAt
     var createdAt by StudentsNotesTable.createdAt
 
     data class New(
         val id: String?,
         val studentId: String,
-        val weight: Double,
-        val waist: Double,
-        val arms: Double,
-        val legs: Double,
+        val measurementName: String,
+        val measurementValue: Double,
+        val tookAt: String,
     )
+
+    data class Page(
+        val id: String,
+        val student: Student.Response,
+        val measurementName: String,
+        val measurementValue: Double,
+        val tookAt: LocalDate,
+    ) {
+        companion object {
+            fun fromDbRow(row: StudentNote): Page {
+                val student = Student.findById(row.studentId)!!
+                return Page(
+                    id = row.id.toString(),
+                    student = Student.Response(
+                        id = row.studentId.toString(),
+                        fullName = student.fullName,
+                        avatar = student.avatar,
+                    ),
+                    measurementName = row.measurementName,
+                    measurementValue = row.measurementValue,
+                    tookAt = row.tookAt,
+                )
+            }
+        }
+    }
 
     data class Response(
         val id: String,
         val studentId: String,
+        val measurementName: String,
+        val measurementValue: Double,
     ) {
         companion object {
             fun fromRow(row: StudentNote): Response = Response(
                 id = row.id.toString(),
                 studentId = row.studentId.toString(),
+                measurementName = row.measurementName,
+                measurementValue = row.measurementValue
             )
         }
     }