+++ test/service/student/StudentMeetingServiceTest.kt
@@ -244,9 +244,4 @@ class StudentMeetingServiceTest {
             assertNotNull(updatedMeeting)
         }
     }
-
-    /*
-
-
-     */
 }